refactor(editor): Register the four auth modals eagerly instead of from the shell (no-changelog)
First consumer of eager, pre-mount registration. All four open from Personal
Settings, which is where the router lands a user whose navigation threw
MfaRequiredError — a path that never reaches the post-login registration in
app/init. Routing them post-login would leave the modal undefined on exactly
that path, silent by the registry gate's design.

Components already lived in the feature; only the keys were shell-owned.
