+++ packages/frontend/editor-ui/src/app/components/ModalRoot.test.ts
@@ -11,11 +11,12 @@ import {
 	CONFIRM_PASSWORD_MODAL_KEY,
 	MFA_SETUP_MODAL_KEY,
 	PROMPT_MFA_CODE_MODAL_KEY,
-} from '@/app/constants';
+} from '@/features/core/auth/auth.constants';
 
 /**
- * The four modal keys `Modals.vue` renders on unauthenticated routes. `<AppModals />`
- * is not auth-gated, so these mount before any registration has run.
+ * The four auth modal keys reachable on unauthenticated routes. `<AppModals />`
+ * is not auth-gated, so these mount before any registration has run — including
+ * before `registerEagerModals()`, which is what now defines them.
  */
 const AUTH_ROUTE_MODAL_KEYS = [
 	CHANGE_PASSWORD_MODAL_KEY,
