+++ packages/frontend/editor-ui/src/app/stores/defaults/modals.ts
@@ -5,9 +5,7 @@ import {
 	AI_BUILDER_DIFF_MODAL_KEY,
 	AI_GATEWAY_TOP_UP_MODAL_KEY,
 	BINARY_DATA_VIEW_MODAL_KEY,
-	CHANGE_PASSWORD_MODAL_KEY,
 	CHAT_EMBED_MODAL_KEY,
-	CONFIRM_PASSWORD_MODAL_KEY,
 	CREDENTIAL_RESOLVER_EDIT_MODAL_KEY,
 	DUPLICATE_MODAL_KEY,
 	EXPERIMENT_TEMPLATE_RECO_V2_KEY,
@@ -19,11 +17,9 @@ import {
 	INSTANCE_AI_CREDENTIAL_SETUP_MODAL_KEY,
 	INSTANCE_AI_TOOLS_CONNECTION_MODAL_KEY,
 	LOG_STREAM_MODAL_KEY,
-	MFA_SETUP_MODAL_KEY,
 	MIGRATE_WORKFLOW_MODAL_KEY,
 	NEW_ASSISTANT_SESSION_MODAL,
 	NPS_SURVEY_MODAL_KEY,
-	PROMPT_MFA_CODE_MODAL_KEY,
 	SECRETS_PROVIDER_CONNECTION_MODAL_KEY,
 	DELETE_SECRETS_PROVIDER_MODAL_KEY,
 	SETUP_CREDENTIALS_MODAL_KEY,
@@ -104,7 +100,6 @@ export const SHELL_MODAL_INITIAL_STATE: Readonly<Record<string, ModalState>> = O
 		},
 	},
 	[BINARY_DATA_VIEW_MODAL_KEY]: { open: false },
-	[CHANGE_PASSWORD_MODAL_KEY]: { open: false },
 	[CHAT_EMBED_MODAL_KEY]: { open: false },
 	[COMMUNITY_PACKAGE_CONFIRM_MODAL_KEY]: {
 		open: false,
@@ -118,7 +113,6 @@ export const SHELL_MODAL_INITIAL_STATE: Readonly<Record<string, ModalState>> = O
 			customHeading: undefined,
 		},
 	},
-	[CONFIRM_PASSWORD_MODAL_KEY]: { open: false },
 	[CREDENTIAL_EDIT_MODAL_KEY]: {
 		open: false,
 		mode: '',
@@ -179,7 +173,6 @@ export const SHELL_MODAL_INITIAL_STATE: Readonly<Record<string, ModalState>> = O
 		open: false,
 		data: undefined,
 	},
-	[MFA_SETUP_MODAL_KEY]: { open: false },
 	[MIGRATE_WORKFLOW_MODAL_KEY]: { open: false },
 	[MOVE_FOLDER_MODAL_KEY]: {
 		open: false,
@@ -192,7 +185,6 @@ export const SHELL_MODAL_INITIAL_STATE: Readonly<Record<string, ModalState>> = O
 	[NPS_SURVEY_MODAL_KEY]: { open: false },
 	[PERSONALIZATION_MODAL_KEY]: { open: false },
 	[PROJECT_MOVE_RESOURCE_MODAL]: { open: false },
-	[PROMPT_MFA_CODE_MODAL_KEY]: { open: false },
 	[SECRETS_PROVIDER_CONNECTION_MODAL_KEY]: { open: false },
 	[SETUP_CREDENTIALS_MODAL_KEY]: { open: false },
 	[SOURCE_CONTROL_PULL_MODAL_KEY]: { open: false },
