+++ packages/frontend/editor-ui/src/features/core/auth/components/ConfirmPasswordModal.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { ref, onMounted } from 'vue';
-import { CONFIRM_PASSWORD_MODAL_KEY } from '@/app/constants';
+import { CONFIRM_PASSWORD_MODAL_KEY } from '../auth.constants';
 import Modal from '@/app/components/Modal.vue';
 import { createFormEventBus } from '@n8n/design-system';
 import type { IFormInputs, IFormInput, FormValues } from '@/Interface';
@@ -8,6 +8,11 @@ import { useI18n } from '@n8n/i18n';
 import { confirmPasswordEventBus } from '../auth.eventBus';
 
 import { N8nButton, N8nFormInputs, N8nText } from '@n8n/design-system';
+
+// Rendered by DynamicModalLoader, which passes the modal-state props every
+// registered modal receives. This one reads its state from its own key, so
+// they must not fall through onto the dialog root as DOM attributes.
+defineOptions({ inheritAttrs: false });
 const config = ref<IFormInputs | null>(null);
 const formBus = createFormEventBus();
 const loading = ref(false);
