+++ packages/frontend/editor-ui/src/features/core/auth/components/PromptMfaCodeModal.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { ref } from 'vue';
 import Modal from '@/app/components/Modal.vue';
-import { PROMPT_MFA_CODE_MODAL_KEY } from '@/app/constants';
+import { PROMPT_MFA_CODE_MODAL_KEY } from '../auth.constants';
 import { useI18n } from '@n8n/i18n';
 import { promptMfaCodeBus } from '../auth.eventBus';
 import { type IFormInput } from '@/Interface';
@@ -9,6 +9,11 @@ import { createFormEventBus } from '@n8n/design-system';
 import { validate as validateUuid } from 'uuid';
 
 import { N8nButton, N8nFormInputs } from '@n8n/design-system';
+
+// Rendered by DynamicModalLoader, which passes the modal-state props every
+// registered modal receives. This one reads its state from its own key, so
+// they must not fall through onto the dialog root as DOM attributes.
+defineOptions({ inheritAttrs: false });
 const i18n = useI18n();
 
 const formBus = createFormEventBus();
